Make Column Filter UI lazy.
+ Fixes mysterious bug.
+ More efficient, lightweight.
