+++ desktop/cmp/grid/impl/filter/ColumnHeaderFilterModel.ts
@@ -6,186 +6,38 @@
  */
 
 import {Column} from '@xh/hoist/cmp/grid';
-import {TabContainerModel} from '@xh/hoist/cmp/tab';
-import {HoistModel, managed} from '@xh/hoist/core';
-import {action, computed, makeObservable, observable} from '@xh/hoist/mobx';
-import {wait} from '@xh/hoist/promise';
+import {HoistModel} from '@xh/hoist/core';
+import {action, makeObservable, observable} from '@xh/hoist/mobx';
 import {isEmpty} from 'lodash';
-import {GridFilterModel, GridFilterFieldSpec} from '@xh/hoist/cmp/grid';
-import {customTab} from './custom/CustomTab';
-import {CustomTabModel} from './custom/CustomTabModel';
-import {valuesTab} from './values/ValuesTab';
-import {ValuesTabModel} from './values/ValuesTabModel';
+import {GridFilterModel} from '@xh/hoist/cmp/grid';
 
 export class ColumnHeaderFilterModel extends HoistModel {
     override xhImpl = true;
 
-    column: Column;
-    gridFilterModel: GridFilterModel;
-    fieldSpec: GridFilterFieldSpec;
+    readonly column: Column;
+    readonly filterModel: GridFilterModel;
 
-    @observable isOpen = false;
-
-    @managed tabContainerModel: TabContainerModel;
-    @managed valuesTabModel: ValuesTabModel;
-    @managed customTabModel: CustomTabModel;
-
-    get field() {
-        return this.fieldSpec.field;
-    }
-
-    get store() {
-        return this.gridFilterModel.gridModel.store;
-    }
-
-    get fieldType() {
-        return this.store.getField(this.field).type;
-    }
-
-    get currentGridFilter() {
-        return this.gridFilterModel.filter;
-    }
-
-    get columnFilters() {
-        return this.gridFilterModel.getColumnFilters(this.field);
-    }
-
-    get columnCompoundFilter() {
-        return this.gridFilterModel.getColumnCompoundFilter(this.field);
-    }
+    @observable isOpen: boolean = false;
 
     get hasFilter() {
-        return !isEmpty(this.columnFilters);
+        const filters = this.filterModel.getColumnFilters(this.column.field);
+        return !isEmpty(filters);
     }
 
-    get hasPendingFilter() {
-        const {activeTabId} = this.tabContainerModel;
-        return activeTabId === 'valuesFilter'
-            ? !!this.valuesTabModel.filter
-            : !!this.customTabModel.filter;
-    }
-
-    @computed
-    get isCustomFilter() {
-        const {columnCompoundFilter, columnFilters} = this;
-        if (columnCompoundFilter) return true;
-        if (isEmpty(columnFilters)) return false;
-        return columnFilters.some(it => !['=', '!=', 'includes'].includes(it.op));
-    }
-
-    get commitOnChange() {
-        return this.gridFilterModel.commitOnChange;
-    }
-
-    constructor({filterModel, column}) {
+    constructor(filterModel: GridFilterModel, column: Column) {
         super();
         makeObservable(this);
-
-        this.gridFilterModel = filterModel;
+        this.filterModel = filterModel;
         this.column = column;
-        this.fieldSpec = filterModel.getFieldSpec(column.field);
-
-        const {enableValues} = this.fieldSpec;
-        this.valuesTabModel = enableValues ? new ValuesTabModel(this) : null;
-        this.customTabModel = new CustomTabModel(this);
-        this.tabContainerModel = new TabContainerModel({
-            switcher: false,
-            tabs: [
-                {
-                    id: 'valuesFilter',
-                    title: 'Values',
-                    content: valuesTab,
-                    omit: !enableValues
-                },
-                {
-                    id: 'customFilter',
-                    title: 'Custom',
-                    content: customTab
-                }
-            ],
-            xhImpl: true
-        });
-
-        this.addReaction({
-            track: () => this.valuesTabModel?.filter,
-            run: () => this.doCommitOnChange('valuesFilter'),
-            debounce: 100
-        });
-
-        this.addReaction({
-            track: () => this.customTabModel.filter,
-            run: () => this.doCommitOnChange('customFilter'),
-            debounce: 100
-        });
     }
 
     @action
-    commit(close = true) {
-        const {tabContainerModel, customTabModel, valuesTabModel} = this,
-            {activeTabId} = tabContainerModel,
-            valuesIsActive = activeTabId === 'valuesFilter',
-            activeTabModel = valuesIsActive ? valuesTabModel : customTabModel,
-            otherTabModel = valuesIsActive ? customTabModel : valuesTabModel;
-
-        this.setColumnFilters(activeTabModel.filter);
-        if (close) {
-            this.closeMenu();
-        } else {
-            // We must wait before resetting as GridFilterModel.setFilter() is async
-            wait().then(() => otherTabModel?.reset());
-        }
-    }
-
-    @action
-    clear(close = true) {
-        this.setColumnFilters(null);
-        if (close) {
-            this.closeMenu();
-        } else {
-            // We must wait before resetting as GridFilterModel.setFilter() is async
-            wait().then(() => this.resetTabModels());
-        }
-    }
-
-    @action
-    openMenu() {
+    open() {
         this.isOpen = true;
-        this.syncWithFilter();
     }
 
     @action
-    closeMenu() {
+    close() {
         this.isOpen = false;
-    }
-
-    //-------------------
-    // Implementation
-    //-------------------
-    @action
-    private syncWithFilter() {
-        const {isCustomFilter, valuesTabModel, customTabModel, tabContainerModel} = this,
-            useCustomTab = isCustomFilter || !valuesTabModel,
-            toTab = useCustomTab ? customTabModel : valuesTabModel,
-            toTabId = useCustomTab ? 'customFilter' : 'valuesFilter';
-
-        this.resetTabModels();
-        toTab.syncWithFilter();
-
-        tabContainerModel.activateTab(toTabId);
-    }
-
-    private setColumnFilters(filters) {
-        this.gridFilterModel.setColumnFilters(this.field, filters);
-    }
-
-    private doCommitOnChange(tab) {
-        if (!this.commitOnChange) return;
-        if (this.tabContainerModel.activeTabId !== tab) return;
-        this.commit(false);
-    }
-
-    private resetTabModels() {
-        this.customTabModel.reset();
-        this.valuesTabModel?.reset();
     }
 }
