+++ desktop/cmp/grid/impl/filter/headerfilter/values/ValuesTabModel.ts
@@ -7,7 +7,7 @@
 import {GridFilterModel, GridModel} from '@xh/hoist/cmp/grid';
 import {HoistModel, managed} from '@xh/hoist/core';
 import {FieldFilterSpec} from '@xh/hoist/data';
-import {ColumnHeaderFilterModel} from '../ColumnHeaderFilterModel';
+import {HeaderFilterModel} from '../HeaderFilterModel';
 import {checkbox} from '@xh/hoist/desktop/cmp/input';
 import {action, bindable, computed, makeObservable, observable} from '@xh/hoist/mobx';
 import {castArray, difference, isEmpty, partition, uniq, without} from 'lodash';
@@ -15,7 +15,7 @@ import {castArray, difference, isEmpty, partition, uniq, without} from 'lodash';
 export class ValuesTabModel extends HoistModel {
     override xhImpl = true;
 
-    headerFilterModel: ColumnHeaderFilterModel;
+    headerFilterModel: HeaderFilterModel;
 
     /** Checkbox grid to display enumerated set of values */
     @managed @observable.ref gridModel: GridModel;
@@ -59,7 +59,7 @@ export class ValuesTabModel extends HoistModel {
     }
 
     get gridFilterModel() {
-        return this.headerFilterModel.gridFilterModel;
+        return this.headerFilterModel.filterModel;
     }
 
     get values() {
@@ -74,7 +74,7 @@ export class ValuesTabModel extends HoistModel {
         return this.values.length < this.valueCount;
     }
 
-    constructor(headerFilterModel: ColumnHeaderFilterModel) {
+    constructor(headerFilterModel: HeaderFilterModel) {
         super();
         makeObservable(this);
 
@@ -83,7 +83,8 @@ export class ValuesTabModel extends HoistModel {
 
         this.addReaction({
             track: () => this.pendingValues,
-            run: () => this.syncGrid()
+            run: () => this.syncGrid(),
+            fireImmediately: true
         });
     }
 
@@ -100,6 +101,7 @@ export class ValuesTabModel extends HoistModel {
     @action
     setRecsChecked(isChecked: boolean, values: any[]) {
         values = castArray(values);
+        console.log(values);
         this.pendingValues = isChecked
             ? uniq([...this.pendingValues, ...values])
             : without(this.pendingValues, ...values);
@@ -193,7 +195,7 @@ export class ValuesTabModel extends HoistModel {
             {fieldType} = headerFilterModel,
             renderer =
                 fieldSpec.renderer ??
-                (fieldType !== 'tags' ? this.headerFilterModel.column.renderer : null);
+                (fieldType !== 'tags' ? this.headerFilterModel.parent.column.renderer : null);
 
         return new GridModel({
             store: {
@@ -224,7 +226,10 @@ export class ValuesTabModel extends HoistModel {
                             disabled: gridModel.store.empty,
                             displayUnsetState: true,
                             value: this.allVisibleRecsChecked,
-                            onChange: () => this.toggleAllRecsChecked()
+                            onChange: () => {
+                                console.log('somebody calling me');
+                                this.toggleAllRecsChecked();
+                            }
                         });
                     },
                     width: 28,
